+++ src/main/java/com/example/jubging/Controller/CommunityController.java
@@ -5,7 +5,7 @@ import com.example.jubging.DTO.PostDTO;
 import com.example.jubging.DTO.PostResultDTO;
 import com.example.jubging.Model.CommunityPost;
 import com.example.jubging.Model.JoinMember;
-import com.example.jubging.Response.SingleResult;
+import com.example.jubging.DTO.Response.SingleResult;
 import com.example.jubging.Service.CommunityService;
 import com.example.jubging.Service.response.ResponseService;
 import lombok.RequiredArgsConstructor;
