+++ src/main/java/com/example/jubging/Controller/PloggingController.java
@@ -3,10 +3,8 @@ package com.example.jubging.Controller;
 import com.example.jubging.DTO.PageDTO;
 import com.example.jubging.DTO.PathwayDTO;
 import com.example.jubging.DTO.RecordDTO;
-import com.example.jubging.Model.Pathway;
-import com.example.jubging.Model.PloggingRecords;
-import com.example.jubging.Response.ListResult;
-import com.example.jubging.Response.SingleResult;
+import com.example.jubging.DTO.Response.ListResult;
+import com.example.jubging.DTO.Response.SingleResult;
 import com.example.jubging.Repository.PloggingRepository;
 import com.example.jubging.Service.RecordService;
 import com.example.jubging.Service.response.ResponseService;
