--- src/main/java/com/example/jubging/DTO/QualificationDTO.java
@@ -1,23 +1,0 @@
-//package com.example.jubging.DTO;
-//
-//import com.example.jubging.Model.CommunityPost;
-//import com.example.jubging.Model.Qualification;
-//import lombok.*;
-//
-//import javax.persistence.Entity;
-//
-//@Getter
-//@Setter
-//@Builder
-//@NoArgsConstructor
-//@AllArgsConstructor
-//public class QualificationDTO {
-//    private String instruction;
-//
-//    public Qualification toEntity(CommunityPost communityPost, String instruction){
-//        return Qualification.builder()
-//                .postId(communityPost)
-//                .instruction(instruction)
-//                .build();
-//    }
-//}
