+++ src/main/java/com/example/jubging/JubgingApplication.java
@@ -1,6 +1,6 @@
 package com.example.jubging;
 
-import com.example.jubging.config.FileStorageProperties;
+import com.example.jubging.common.config.FileStorageProperties;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
