+++ src/main/java/com/example/jubging/Repository/UserRepository.java
@@ -14,4 +14,5 @@ public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findById(Long id);
     boolean existsByUserId(String userId);
     boolean existsByNickname(String nickname);
+    Optional<User> findByUserIdOrnOrNickname(String userId, String nickname);
 }
