+++ src/main/java/com/example/jubging/Service/CommunityService.java
@@ -1,8 +1,8 @@
 package com.example.jubging.Service;
 
 import com.example.jubging.DTO.*;
-import com.example.jubging.Exception.CUserNotFoundException;
-import com.example.jubging.Exception.CommunityCapacityException;
+import com.example.jubging.common.Exception.CUserNotFoundException;
+import com.example.jubging.common.Exception.CommunityCapacityException;
 import com.example.jubging.Model.CommunityPost;
 import com.example.jubging.Model.JoinMember;
 import com.example.jubging.Model.Qualification;
@@ -11,10 +11,9 @@ import com.example.jubging.Repository.CommunityPostingRepository;
 import com.example.jubging.Repository.JoinMemberRepository;
 import com.example.jubging.Repository.QualificationRepository;
 import com.example.jubging.Repository.UserRepository;
-import com.example.jubging.config.security.JwtTokenProvider;
+import com.example.jubging.auth.JwtTokenProvider;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.hibernate.mapping.Join;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
@@ -22,9 +21,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.servlet.http.HttpServletRequest;
-import java.sql.Date;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
