+++ src/main/java/com/example/jubging/Service/RecordService.java
@@ -2,17 +2,16 @@ package com.example.jubging.Service;
 
 import com.example.jubging.DTO.PageDTO;
 import com.example.jubging.DTO.PathwayDTO;
-import com.example.jubging.Exception.CEmailLoginFailedException;
+import com.example.jubging.common.Exception.CEmailLoginFailedException;
 import com.example.jubging.DTO.RecordDTO;
-import com.example.jubging.Exception.CUserNotFoundException;
-import com.example.jubging.Model.Pathway;
+import com.example.jubging.common.Exception.CUserNotFoundException;
 import com.example.jubging.Model.Pathway;
 import com.example.jubging.Model.PloggingRecords;
 import com.example.jubging.Model.User;
 import com.example.jubging.Repository.PathwayRepository;
 import com.example.jubging.Repository.PloggingRepository;
 import com.example.jubging.Repository.UserRepository;
-import com.example.jubging.config.security.JwtTokenProvider;
+import com.example.jubging.auth.JwtTokenProvider;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
@@ -23,7 +22,6 @@ import org.springframework.transaction.annotation.Transactional;
 
 import javax.servlet.http.HttpServletRequest;
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Slf4j
