+++ src/main/java/com/example/jubging/common/Exception/CAuthenticationEntryPointException.java
@@ -1,4 +1,4 @@
-package com.example.jubging.Exception;
+package com.example.jubging.common.Exception;
 
 public class CAuthenticationEntryPointException extends RuntimeException{
     public CAuthenticationEntryPointException() {
