+++ src/main/java/com/example/jubging/common/Exception/CUserNotFoundException.java
@@ -1,4 +1,4 @@
-package com.example.jubging.Exception;
+package com.example.jubging.common.Exception;
 
 public class CUserNotFoundException extends RuntimeException{
