+++ src/main/java/com/example/jubging/common/Exception/CommunityCapacityException.java
@@ -1,4 +1,4 @@
-package com.example.jubging.Exception;
+package com.example.jubging.common.Exception;
 
 public class CommunityCapacityException extends RuntimeException{
     public CommunityCapacityException (){super();}
