+++ src/main/java/com/example/jubging/common/Exception/FileStorageException.java
@@ -1,4 +1,4 @@
-package com.example.jubging.Exception;
+package com.example.jubging.common.Exception;
 
 public class FileStorageException extends RuntimeException{
     public FileStorageException(String message) {
