+++ src/main/java/com/example/jubging/common/config/OpenApiConfig.java
@@ -1,4 +1,4 @@
-package com.example.jubging.config;
+package com.example.jubging.common.config;
 
 import io.swagger.v3.oas.annotations.OpenAPIDefinition;
 import io.swagger.v3.oas.annotations.info.Contact;
