+++ src/main/resources/application.yml
@@ -3,5 +3,5 @@ server:
 
 spring:
   profiles:
-    active: email
+    active: prod
     include: jwt
